medtheris-scraper/requirements.txt
googlemaps>=4.10
playwright>=1.49
httpx>=0.27
beautifulsoup4>=4.12
anthropic>=0.40
gql[requests]>=3.5
python-dotenv>=1.0
